+++ design-system/badge/src/components/__tests__/unit/index.tsx
@@ -1,24 +1,93 @@
 import React from 'react';
 
-import { mount } from 'enzyme';
+import { render } from '@testing-library/react';
 
-import Format from '../../Format';
 import Badge from '../../index';
 
-function getFormatProps(badge: React.ReactElement<any>) {
-  return mount(badge).find(Format).props();
-}
+describe('badge component', () => {
+  const testId = 'test';
 
-test('snapshot', () => {
-  expect(mount(<Badge />)).toMatchSnapshot();
-});
+  it('should render 0 by default', () => {
+    const { getByText } = render(<Badge />);
+    expect(getByText('0')).toBeInTheDocument();
+  });
 
-test('children', () => {
-  expect(getFormatProps(<Badge>{100}</Badge>)).toMatchObject({ children: 100 });
-});
+  it.each([0, 100, 12.34])(
+    'should render positive numeric children (value=%p)',
+    value => {
+      const { getByText } = render(<Badge max={value}>{value}</Badge>);
+      expect(getByText(value.toString())).toBeInTheDocument();
+    },
+  );
 
-test('using a string', () => {
-  expect(mount(<Badge>+100</Badge>).props()).toMatchObject({
-    children: '+100',
+  it.each([-1, -100, -Infinity])(
+    'should clamp negative numeric children (value=%p)',
+    value => {
+      const { getByText } = render(<Badge>{value}</Badge>);
+      expect(getByText('0')).toBeInTheDocument();
+    },
+  );
+
+  it.each(['-100', '0', '100', 'abc', '+100,000.333'])(
+    'should render string children exactly (value=%p)',
+    value => {
+      const { getByText } = render(<Badge>{value}</Badge>);
+      expect(getByText(value)).toBeInTheDocument();
+    },
+  );
+
+  it('should have max=99 by default', () => {
+    const { getByText } = render(<Badge>{100}</Badge>);
+    expect(getByText('99+')).toBeInTheDocument();
+  });
+
+  it.each([
+    [10, 100, '10'],
+    [1000, 100, '100+'],
+  ])(
+    'should respect positive values of max (value=%p, max=%p, expected=%p)',
+    (value, max, expected) => {
+      const { getByText } = render(<Badge max={max}>{value}</Badge>);
+      expect(getByText(expected)).toBeInTheDocument();
+    },
+  );
+
+  it.each([
+    [0, -1, '0'],
+    [100, 0, '100'],
+    [Infinity, -100, '∞'],
+  ])(
+    'should ignore non-positive values for max (value=%p, max=%p, expected=%p)',
+    (value, max, expected) => {
+      const { getByText } = render(<Badge max={max}>{value}</Badge>);
+      expect(getByText(expected)).toBeInTheDocument();
+    },
+  );
+
+  it.each([0, -100, -Infinity])(
+    'should clamp negative numeric values (value=%p, expected="0")',
+    value => {
+      const { getByText } = render(<Badge>{value}</Badge>);
+      expect(getByText('0')).toBeInTheDocument();
+    },
+  );
+
+  it.each([
+    [Infinity, -100, '∞'],
+    [1000, Infinity, '1000'],
+    [Infinity, 100, '100+'],
+    [Infinity, Infinity, '∞'],
+  ])(
+    'should handle Infinity (value=%p, max=%p, expected=%p)',
+    (value, max, expected) => {
+      const { getByText } = render(<Badge max={max}>{value}</Badge>);
+      expect(getByText(expected)).toBeInTheDocument();
+    },
+  );
+
+  it('should render with a given test id', () => {
+    const { getByTestId } = render(<Badge testId={testId} />);
+    const element = getByTestId(testId);
+    expect(element).toBeInTheDocument();
   });
 });